+++ django_peeringdb/client_adaptor/__init__.py
@@ -1,1 +1,0 @@
-from django_peeringdb.client_adaptor.load import load_backend