+++ src/django_peeringdb/__init__.py
@@ -1,0 +1,8 @@
+try:
+    import importlib.metadata as importlib_metadata
+# python < 3.8
+except ImportError:
+    import importlib_metadata
+
+
+__version__ = importlib_metadata.version(__name__)