+++ src/django_peeringdb/client_adaptor/__init__.py
@@ -1,0 +1,1 @@
+from django_peeringdb.client_adaptor.load import load_backend  # noqa