+++ tests/peeringdb_mock/backend.py
@@ -7,15 +7,15 @@ def reftag_to_cls(fn):
     return fn
 
 
-class Resource(object):
+class Resource:
     def __init__(self, tag):
         self.tag = tag
 
 
-class Interface(object):
-    REFTAG_RESOURCE = dict(
-        [(model.HandleRef.tag, Resource(model.HandleRef.tag)) for model in all_models]
-    )
+class Interface:
+    REFTAG_RESOURCE = {
+        model.HandleRef.tag: Resource(model.HandleRef.tag) for model in all_models
+    }
 
     def get_resource(self, concrete):
         return self.REFTAG_RESOURCE[concrete.HandleRef.tag]